The maximum for the resource limits has been changed from ulong.MaxValue to long.MaxValue in ImageMagick.

+++ Tests/Magick.NET.Tests/Shared/ResourceLimitsTests/TheListLengthProperty.cs
@@ -23,7 +23,7 @@ namespace Magick.NET.Tests
             [TestMethod]
             public void ShouldHaveTheCorrectValue()
             {
-                Assert.AreEqual(ulong.MaxValue, ResourceLimits.ListLength);
+                Assert.AreEqual((ulong)long.MaxValue, ResourceLimits.ListLength);
             }
 
             [TestMethod]
